cases/gemini/requirements.txt
google-genai
python-dotenv